+++ CHANGELOG.md
@@ -48,5 +48,6 @@ This update adds 4 new items (3 are Shield-oriented) and 2 equipment, as well as
 
 ## v1.2.0
 This update adds more alternate skills to a few characters.
+* Added skill Stealth Hunting (Huntress special)
 * Added skill Unload (Bandit secondary)
 * Magic Dagger : Debuff no longer remains when an enemy becomes unstunned early.